Add ring buffer class
Asynchronous IO is pretty related with ring buffers. For
stream-oriented protocols where message size cannot be determined
before read, it is pretty important to have the capability to read
as many bytes as available until a new message can be deserialized.
After that, there could be remaining bytes that belong to the next
message. Those bytes have to be retained, and there is where the
ring buffer comes into action.

This change adds a new StreamBuffer class named ring_buffer, and
adapts the asynchronous TCP server to use StreamBuffer, allowing
a client to use a ring_buffer for stream-oriented communications.