Introduce a new FV entity: variable ID
There are several parts in the code where the pair of var group
and var name are used. In order to reduce this boilerplate code,
a new entity is introduced: the variable_id type. This is just a
std::pair or var group and name, with convenience functions to
handle it. In a future commit, the API will be refactored to use
variable_id instead of splitted var group and name arguments.