Fix some building issues
Issues related with WinSocket.h and Windows.h includes appear again
in Boost 1.54.0. This commit defines the macro WIN32_LEAN_AND_MEAN
before every include statement for Windows.h header, finally fixing
the issue.

FSUIPC SDK building is a real pain. The latest version that may be
found in the website lacks the MSVC project file to regenerate the
library. So a extreme decision has been made for that: include the
FSUIPC user code into OACSD codebase. A new fsuipc subproject is
added, which contains the source code of FSUIPC User Internal with
an additional copyright note indicating who are the owners of these
files. This is fully compliant with both the GPL license and the
public domain status of FSUIPC SDK. We may expect not to have any
further issue with this dependency anymore.