Add some amazing utilities to FSUIPC
The buffer-oriented nature of FSUIPC abstraction provided by liboac
is not the best suitable. It's not trivial to observe updates in
offsets due to this buffer orientation.

This change provides new entities to integrate user code with FSUIPC.

 - A fsuipc_client, which abstracts the FSUIPC read and writes in
   terms of query() and update() operations. This functions operate
   on collections of offfsets rather than single ones, so it's much
   more appropriate for read-many, write-many loops. This client
   is decoupled from the FSUIPC API being used via a the new concept
   FsuipcUserAdapter.

 - A fsuipc_update_observer, which wraps a fsuipc_client object and
   facilitates the observation of changes on the offsets indicated
   by client code.

A new unit test source was added, which name (fsuipc-test) clashed
other one provided in flight-vars. This has been fixed by refactorign
cmake's add_unit_test() function to admit a new argument to be used
as prefix for the targets.