Fix a dereferentiation of a deallocated pointer
There was a nasty bug in thread_worker that was causing a potential
dereferentiation of a deallocated pointer. The new thread was created with
a lambda function as start point, capturing 'this' variable. It was
possible that the thread_worker was deallocated before the new thread
began to execute. When it starts, it dereferenciates the 'this' pointer
causing a potential memory corruption or SIGSEGV.