Extract io_service from async_tcp_server
So far, the `boost::asio::io_service` used by `async_tcp_server` was
initialized by the constructor of the latter. That prevents the IO service
object to be used for other purposes.

This change refactors `async_tcp_server` so it accepts a `io_service`
object in its constructor. In addition, the lifecycle of the IO service
and the thread executing its event loop is decoupled from the async
server.

Also some clean up of the network.h file was done, including the coverage
of connection errors in `tcp_client` and the corresponding tests in
`network-test.cpp`.