Adapt server-itest for external IO service
This change adapts server-itest in order to externalize the IO service
from the flight_vars_server object. With the io_service instance
externalized, it's possible to test how fsuipc_flight_vars can be checked
for updates using the IO service and the same thread that is serving the
TCP connections.

This closes the issue #12.