Add a Boost ASIO adapter for TickObserver
As issue #16 stands, a class that adapts a TickObserver in order to
handlers to be dispatched via Boost IO service is required. This change
adds a new class asio_tick_observer_adapter for that purpose.

In addition, a new class concurrent_tick_observer_base is added to support
the implementation of TickObserver objects where registration and
notification of ticks may be executed concurrently (as the case of
SimConnect tick observer).