Add log author support in liboac-commons
As issue #10 explains, the logging facilities provided by liboac-commons
doesn't allow to specify the module name or author of each log entry.

This change modifies the `log_*`()` functions to require an additional
argument to indicate the author of the entry. In addition, a new class
`logger_component` is provided, which receives such an author in its
constructor and provides a rich set of convenience protected functions for
logging.

Of course, the log actions all over the code has been modified to adequate
to this new interface.