Fix a crash in FlightVars TCP server
When TCP server is waiting for a begin session message, it may receive
garbage (bytes chunk that cannot be recognized as a message). In that
case, an exception is thrown but uncaught, which causes the process to
terminate.

This commit fixes this issue by adding a catch statement in
`on_read_begin_session()`. In addition, some new test cases have been
added to check how the server responds against chunks of garbage.