Split TCP server thread from tick observer
So far, the same thread is attending both the TCP server and the tick
observer which is performing blocking actions on FSUIPC FlightVars
instance. See issue #19 for further details.

This change forces `flight_vars_server` to execute the var update handler
in the same thread that is attending the TCP connections. Along this, the
plugin main's function is creating a `simconnect_tick_observer` that may
be used to check for updates in `fsuipc_flight_vars` object.