Fix random failures in server-itest.cpp
In FligtVars integration test, when garbage is sent to the server the test
case expects the connection to be reset by peer. Sometimes the check fails
due to the server has not processed yet the message that causes the
failure.

This change makes the thread to sleep a while before checking the
connection is reset.