Fix a bug in linear_buffer
The callback functions `on_async_read` and `on_async_write` of
`linear_buffer` class were wrongly updating the state of the object. In
this context, and async read must increment the *write* index, and
async write must increment the *read* index. The code was doing the other
way around, causing the buffer not to work when used for async operations.