Fix broken test case in server-itest.cpp
There was a broken test case in `server-itest.cpp` caused by
synchronization problem. The server was unable to process an invalid
message sent by the client in handshake phase in the given time. The
period the client thread is sleeping is incremented to let the server
reject the message, close the connection and the client side to be aware
of all this.