Make FSUIPC update as soon as var is subscribed
As reported in issue #46, it is essential that client is notified with a
variable update once a subscription is successfully received by the
server.

This change applies the changes needed to `update_observer` to obtain such
a behaviour. Tests are updated as well to take into account such a first
variable update immediately after each subscription request.