Make installer choose the simulator
As reported in issue #47, the user should be able to choose the simulator
he wants to install OAC plugins for. This change cleans up the NSIS script
by moving the plugin handling code to `fs-plugins.nsh`. A new custom page
is added as well to choose the simulator that will be configured to use
the plugins to be installed.