Fix a bug in `mosquitto_client`
A client that doesn't succeed to connect to the broker is left in a
inconsistent state. The mosquitto handler is not destroyed, and
therefore the connect messages may be still travelling all around. If
a new broker appears that is able to response to the connect messages,
the connect handler will be invoked on an already destroyed client
object.

This commit fixes the bug, but a more robust approach could be
implemented using a FSM to track the client state and avoid
unnecessary or illegal actions to be executed in destruction.