Integrate all tests into a single executable
One executable per test suite is a explosion of binaries that causes
long term compilations and a lot of subprojects to be generated in the
IDE.

This change integrates all the tests in two executables: one for
unit test and other one for integration tests. This applies to liboac
project. More projects coming next.