Add some util functions to use WIN32 API
WIN32 API functions use to return integer values indicating whether
the things are OK and, otherwise, a `GetLastError()` function is
provided to check what's the exact cause of the problem. While coding
in C++, a terrible pattern is repeated again and again: invoke the
function, check the result, if failure invoke `GetLastError()` and
then throw the exception.

This change provides a util header `liboac/util/win.h` that provides
utility functions and macros to invoke WIN32 API functions that
converts their error codes into an exception.