Replace Mosquitto service runner by none in config
Having a Mosquitto Windows service MQTT broker runner is not trivial.
On Windows platforms, a process must have special privileges to
manage system services (makes sense after all). Therefore, it is not
possible that FlightVars plugin manages the state of Mosquitto broker
service.

New approach: after Mosquitto installation, the broker service is
installed and configured to be started on each reboot. So there
should be a running broker on any windows box with Mosquitto installed.
That means... no need for a runner!

This change removes the `MOSQUITTO_SERVICE` from the
`mqtt_broker_runner_id` enum and provides a new default one: `NONE`.