Some refactors in MQTT client
The subscription interface of `mqtt::client` is not appropriate for
the purpose of FlightVars. The subscription callback was designed to
push the message data to the subscriber, ommiting any other relevant
information.

This change refactors the interface of `mqtt::client` with the
incorporation of a new operation `subscribe_as_message()` that accepts
a generic message as argument. That generic message contains the topic,
QoS, retain flag and data.