Remove MQTT client IT dependency to its runner
MQTT client IT used to depend on a MQTT broker runner in order to
control the presence or absence of the broker. This was due to the
necessity to test the client against a unavailable broker, but caused
problems to run the process when the Mosquitto service was already
running.

Now the IT depends exclusively on a running Mosquitto service. The
unavailable broker is simulated by a client attempting to connect to
a non-existing service in a remote host.