Provide encapsulation on FSUIPC client objects
The previous design of FSUIPC client objects were based upon the idea
of having getters for collaboration objects that make possible to
access their internal state. E.g., `fsuipc_client` instantiates its own
`FsuipcUserAdapter` object and provides a getter to retrieve it. This
is a violation of the encapsulation principle.

This change modifies this set of objects in order to receive the
collaborator as shared objects in their constructors. Therefore, it is
possible to pass a certain `FsuipcUserAdapter` object to `fsuipc_client`
while the adapter is kept by the creator in order to check its internal
state if desired.