Fix a compilation error caused by ERROR macro
Somewhere in the Win32 API there is a macro declaration for ERROR.
This macro, again, is clashing with the value ERROR of the `log_level`
enum.

Fucking stupid Microsoft engineers: do you really though nobody would
be interested on using such a token in their programs? I've lost half
an hour chasing the source of this stupid error. Thank you very much
for the frustrating experience that represents programming in this
bullshit of platform.