Fix the FSUIPC_Process() never returns bug
`FSUIPC_Process()` is calling WIN32's `SendMessage()`, which blocks
the caller thread until the message is processed. That's great until
FSUIPC is not receiving messages anymore (e.g., when FSX is closing
and FSUIPC plugin has been stopped).

This patch replaces `SendMessage()` by `SendMessageTimeout()` to return
with an error in case of FSUIPC is not responding to the message.