Move thread machinery from module to domain
The threading machinery (the task executor and the recurrent task) that
runs the `fsuipc::domain` logic was instantiated and managed in the
`module` class. That's incorrect since it means race conditions in
the domain object (MQTT client thread vs task executor thread).

This change moves all threading machinery into the `fsuipc::domain`.
Now the callback for MQTT messages, instead of executing the appropriate
action, creates a task in the executor for it to do that. No more race
conditions in the domain object.