Changes in FSUIPC Offsets message format
So far the FSUIPC Offsets domain used a fixed 32-bits length message
to propagate the offset values. That represents two drawbacks:
* There is no reason to send 4 bytes for offsets of 1 or 2 bytes length.
* There is no indication to determine what's the sender of the message.
Therefore, it is impossible for FV to filter the messages that has been
sent by itself.

This change replaces the 32-bits number by a message structure that
addresses these two issues.