Use custom loop rather than Mosquitto one
The function `mosquitto_loop_start` is a fucking bullshit. For any
reason I don't know, it queues the outcoming publish messages for
almost a second. That introduces a big lag in the messages sent by
flightvars, causing a really bad user experience.

The internal thread provided by Mosquitto is not used in this patch.
Instead, a custom thread is used. This definitively addresses the
problem.